+++ tests/__init__.py
@@ -17,6 +17,12 @@ from scrapy_zyte_api.utils import _POET_ADDON_SUPPORT
 
 _API_KEY = "a"
 
+DEFAULT_AUTOMAP_PARAMS: Dict[str, Any] = {
+    "httpResponseBody": True,
+    "httpResponseHeaders": True,
+    "responseCookies": True,
+}
+
 DEFAULT_CLIENT_CONCURRENCY = AsyncClient(api_key=_API_KEY).n_conn
 SETTINGS_T = Dict[str, Any]
 SETTINGS: SETTINGS_T = {