+++ tests/test_middlewares.py
@@ -112,8 +112,7 @@ async def test_preserve_delay(mw_cls, processor, settings, preserve):
 async def test_cookies():
     """Make sure that the downloader middleware does not crash on Zyte API
     requests with cookies."""
-    settings = {"ZYTE_API_EXPERIMENTAL_COOKIES_ENABLED": True}
-    crawler = get_crawler(settings_dict=settings)
+    crawler = get_crawler()
     await crawler.crawl("a")
     spider = crawler.spider
     middleware = create_instance(