+++ tests/test_request_fingerprinter.py
@@ -108,6 +108,69 @@ async def test_headers():
     assert fingerprint1 == fingerprint2
 
 
+@ensureDeferred
+async def test_cookies():
+    crawler = await get_crawler()
+    fingerprinter = create_instance(
+        ScrapyZyteAPIRequestFingerprinter, settings=crawler.settings, crawler=crawler
+    )
+    request1 = Request(
+        "https://example.com",
+        meta={
+            "zyte_api": {
+                "responseCookies": False,
+                "requestCookies": [{"name": "foo", "value": "bar"}],
+                "cookieManagement": False,
+                "experimental": {
+                    "responseCookies": False,
+                    "requestCookies": [{"name": "foo", "value": "bar"}],
+                    "cookieManagement": False,
+                },
+            }
+        },
+    )
+    # Same with responseCookies set to `True`.
+    request2 = Request(
+        "https://example.com",
+        meta={
+            "zyte_api": {
+                "responseCookies": True,
+                "requestCookies": [{"name": "foo", "value": "bar"}],
+                "cookieManagement": False,
+                "experimental": {
+                    "responseCookies": False,
+                    "requestCookies": [{"name": "foo", "value": "bar"}],
+                    "cookieManagement": False,
+                },
+            }
+        },
+    )
+    # Same with experimental.responseCookies set to `True`.
+    request3 = Request(
+        "https://example.com",
+        meta={
+            "zyte_api": {
+                "requestCookies": [{"name": "foo", "value": "bar"}],
+                "cookieManagement": False,
+                "experimental": {
+                    "responseCookies": True,
+                    "requestCookies": [{"name": "foo", "value": "bar"}],
+                    "cookieManagement": False,
+                },
+            }
+        },
+    )
+    request4 = Request("https://example.com", meta={"zyte_api": True})
+    fingerprint1 = fingerprinter.fingerprint(request1)
+    fingerprint2 = fingerprinter.fingerprint(request2)
+    fingerprint3 = fingerprinter.fingerprint(request3)
+    fingerprint4 = fingerprinter.fingerprint(request4)
+    assert fingerprint1 != fingerprint2
+    assert fingerprint1 != fingerprint3
+    assert fingerprint1 == fingerprint4
+    assert fingerprint2 == fingerprint3
+
+
 @pytest.mark.parametrize(
     "url,params,fingerprint",
     (
@@ -263,9 +326,7 @@ async def test_only_end_parameters_matter():
             "zyte_api": {
                 "httpResponseBody": True,
                 "httpResponseHeaders": True,
-                "experimental": {
-                    "responseCookies": True,
-                },
+                "responseCookies": True,
             }
         },
     )